pyproject.toml
[tool.poetry]
name = "investment-automation"
version = "0.0.1"
description = ""
authors = ["Aaron Mamparo <aaronmamparo@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4"
python-dotenv = "^1.0.0"
injector = "^0.20.1"
boto3 = "^1.28.1"
requests = "2.29.0"
pandas = "^2.0.3"
scipy = "^1.11.1"
playwright = "^1.45.0"
pytest-playwright = "^0.5.1"

[tool.poetry.group.dev.dependencies]
aws-cdk-lib = "^2.87.0"
pylint = "^2.17.4"
pylint-quotes = "^0.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
